requirements_web.txt
google-cloud-vision
Pillow
opencv-python
streamlit
flask
gunicorn
